# apps/server/openapi.yaml
openapi: 3.0.3
info:
  title: OpenBTS API
  description: API for OpenBTS Backend
  version: 1.0.0
servers:
  - url: https://openbts.sakilabs.com/api/v1
    description: API v1 base path
...
paths:
...
  /cells:
    get:
      tags: [Cells]
      summary: List cells
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 150 }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/CellWithDetailsAndBand" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      tags: [Cells]
      summary: Create cell
      description: In `details`, do not send **cid_long** (UMTS), **ecid** (LTE) or **nci** (NR); they are computed on the server.
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CellCreate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/CellWithDetails" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /stations/{station_id}/photos/{photo_id}:
    patch:
      tags: [Stations]
      summary: Update a station photo (set as main or update note)
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      parameters:
        - $ref: "#/components/parameters/StationIdPath"
        - name: photo_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_main: { type: boolean, enum: [true], description: "Set this photo as the main photo" }
                note: { type: string, maxLength: 100, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [id, is_main, note]
                    properties:
                      id: { type: integer }
                      is_main: { type: boolean }
                      note: { type: string, nullable: true }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: [Stations]
      summary: Delete a station photo
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      parameters:
        - $ref: "#/components/parameters/StationIdPath"
        - name: photo_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /submissions/{id}/photos/{photo_id}:
    patch:
      tags: [Submissions]
      summary: Update a submission photo note
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: photo_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                note: { type: string, maxLength: 100 }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: [Submissions]
      summary: Delete a submission photo
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: photo_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API Key for authentication via X-Api-Key header
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: Cookie-based authentication
  parameters:
...
    StationIdPath:
      name: station_id
      in: path
      required: true
      schema: { type: integer }
  responses:
    BadRequest:
      description: Bad request (invalid parameters, validation error, or invalid query)
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [BAD_REQUEST, VALIDATION_ERROR, INVALID_QUERY]
          example:
            errors:
              - code: BAD_REQUEST
                message: Invalid request
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [UNAUTHORIZED]
          example:
            errors:
              - code: UNAUTHORIZED
                message: Unauthorized
    Forbidden:
      description: Forbidden or insufficient permissions
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [FORBIDDEN, INSUFFICIENT_PERMISSIONS]
          example:
            errors:
              - code: FORBIDDEN
                message: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [NOT_FOUND]
          example:
            errors:
              - code: NOT_FOUND
                message: Resource not found
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [TOO_MANY_REQUESTS]
          example:
            errors:
              - code: TOO_MANY_REQUESTS
                message: Too many requests
    Conflict:
      description: Conflict (e.g. duplicate entry)
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [DUPLICATE_ENTRY]
          example:
            errors:
              - code: DUPLICATE_ENTRY
                message: Duplicate entry
    InternalServerError:
      description: Internal server error (e.g. failed to create/update/delete)
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [INTERNAL_SERVER_ERROR, FAILED_TO_CREATE, FAILED_TO_UPDATE, FAILED_TO_DELETE, AUTH_FAILURE]
          example:
            errors:
              - code: INTERNAL_SERVER_ERROR
                message: Internal server error
  schemas:
    ApiError:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required: [code, message]
            properties:
              code:
                type: string
                enum:
                  - INTERNAL_SERVER_ERROR
                  - BAD_REQUEST
                  - NOT_FOUND
                  - UNAUTHORIZED
                  - FORBIDDEN
                  - FAILED_TO_DELETE
                  - INVALID_QUERY
                  - FAILED_TO_UPDATE
                  - FAILED_TO_CREATE
                  - AUTH_FAILURE
                  - VALIDATION_ERROR
                  - ALREADY_LOGGED_IN
                  - INSUFFICIENT_PERMISSIONS
                  - TOO_MANY_REQUESTS
                  - DUPLICATE_ENTRY
              message: { type: string }
              details:
                type: array
                description: Only present for VALIDATION_ERROR
                items:
                  type: object
                  properties:
                    field: { type: string }
                    validationMessage: { type: string, nullable: true }
...
    Band:
      type: object
      properties:
        id: { type: integer }
        value: { type: integer, nullable: true }
        rat: { type: string, enum: [GSM, CDMA, UMTS, LTE, NR, IOT] }
        name: { type: string }
        duplex: { type: string, enum: [FDD, TDD], nullable: true }
        variant: { type: string, enum: [commercial, railway] }
...
    Operator:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        full_name: { type: string }
        parent_id: { type: integer, nullable: true }
        mnc: { type: integer, nullable: true }
...
    Location:
      type: object
      properties:
        id: { type: integer }
        city: { type: string, nullable: true }
        address: { type: string, nullable: true }
        longitude: { type: number }
        latitude: { type: number }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...
    Station:
      type: object
      properties:
        id: { type: integer }
        station_id: { type: string }
        notes: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        is_confirmed: { type: boolean, nullable: true }
        operator: { $ref: "#/components/schemas/Operator", nullable: true }
        location: { $ref: "#/components/schemas/Location", nullable: true }
        extra_identificators: { $ref: "#/components/schemas/ExtraIdentificators", nullable: true }
...
    CellBase:
      type: object
      properties:
        id: { type: integer }
        station_id: { type: integer }
        band_id: { type: integer }
        rat: { type: string, enum: [GSM, CDMA, UMTS, LTE, NR, IOT] }
        notes: { type: string, nullable: true }
        is_confirmed: { type: boolean, nullable: true }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    GsmCellDetails:
      type: object
      properties:
        lac: { type: integer }
        cid: { type: integer }
        e_gsm: { type: boolean, nullable: true }
    UmtsCellDetails:
      type: object
      properties:
        lac: { type: integer, nullable: true }
        arfcn: { type: integer, nullable: true }
        rnc: { type: integer }
        cid: { type: integer }
        cid_long: { type: integer }
    LteCellDetails:
      type: object
      properties:
        tac: { type: integer, nullable: true }
        enbid: { type: integer }
        clid: { type: integer }
        ecid: { type: integer }
        supports_nb_iot: { type: boolean, nullable: true }
    NrCellDetails:
      type: object
      properties:
        nrtac: { type: integer, nullable: true }
        gnbid: { type: integer, nullable: true }
        gnbid_length: { type: integer, nullable: true }
        clid: { type: integer, nullable: true }
        nci: { type: integer, nullable: true }
        pci: { type: integer, nullable: true }
        supports_nr_redcap: { type: boolean, nullable: true }
    UmtsCellDetailsInsert:
      description: UMTS cell details for create/update. Do not send cid_long; it is computed server-side from rnc and cid.
      type: object
      properties:
        lac: { type: integer, nullable: true }
        arfcn: { type: integer, nullable: true }
        rnc: { type: integer }
        cid: { type: integer }
    LteCellDetailsInsert:
      description: LTE cell details for create/update. Do not send ecid; it is computed server-side from enbid and clid.
      type: object
      properties:
        tac: { type: integer, nullable: true }
        enbid: { type: integer }
        clid: { type: integer }
        supports_nb_iot: { type: boolean, nullable: true }
    NrCellDetailsInsert:
      description: NR cell details for create/update. Do not send nci; it is computed server-side from gnbid, gnbid_length and clid.
      type: object
      properties:
        nrtac: { type: integer, nullable: true }
        gnbid: { type: integer, nullable: true }
        gnbid_length: { type: integer, nullable: true }
        clid: { type: integer, nullable: true }
        pci: { type: integer, nullable: true }
        supports_nr_redcap: { type: boolean, nullable: true }
    CellDetailsInsert:
      description: Cell RAT-specific payload for create/update. Omit server-calculated fields (cid_long for UMTS, ecid for LTE, nci for NR).
      oneOf:
        - $ref: "#/components/schemas/GsmCellDetails"
        - $ref: "#/components/schemas/UmtsCellDetailsInsert"
        - $ref: "#/components/schemas/LteCellDetailsInsert"
        - $ref: "#/components/schemas/NrCellDetailsInsert"
    CellDetails:
      oneOf:
        - $ref: "#/components/schemas/GsmCellDetails"
        - $ref: "#/components/schemas/UmtsCellDetails"
        - $ref: "#/components/schemas/LteCellDetails"
        - $ref: "#/components/schemas/NrCellDetails"
    CellWithDetails:
      allOf:
        - $ref: "#/components/schemas/CellBase"
        - type: object
          properties:
            station: { $ref: "#/components/schemas/Station", nullable: true }
            band: { $ref: "#/components/schemas/Band", nullable: true }
            details: { $ref: "#/components/schemas/CellDetails", nullable: true }
    CellWithDetailsAndBand:
      allOf:
        - $ref: "#/components/schemas/CellBase"
        - type: object
          properties:
            band: { $ref: "#/components/schemas/Band" }
            details: { $ref: "#/components/schemas/CellDetails", nullable: true }
...
    CellCreate:
      description: Do not send cid_long (UMTS), ecid (LTE) or nci (NR) in details; they are computed on the server.
      type: object
      required: [station_id, band_id, rat]
      properties:
        station_id: { type: integer }
        band_id: { type: integer }
        rat: { type: string, enum: [GSM, CDMA, UMTS, LTE, NR, IOT] }
        notes: { type: string, nullable: true }
        is_confirmed: { type: boolean, nullable: true }
        details: { $ref: "#/components/schemas/CellDetailsInsert" }
...
    ExtraIdentificators:
      type: object
      properties:
        id: { type: integer }
        networks_id: { type: integer, nullable: true }
        networks_name: { type: string, nullable: true }
        mno_name: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...